+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add support for Ruby 3.3
 - Allow SyncProcessor to be called from appenders
 - Fix incorrect metrics usage examples in documentation
+- Add `:duration_ms` to Logfmt fomatter
 
 ## [4.15.0]